requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.8.2
captcha==0.7.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
Deprecated==1.2.18
Flask==3.0.3
flask-cors==5.0.1
Flask-Limiter==3.10.1
itsdangerous==2.2.0
Jinja2==3.1.4
libsass==0.23.0
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
peewee==3.17.6
pillow==11.1.0
pycparser==2.22
Pygments==2.19.1
python-dotenv==1.0.1
rich==13.9.4
typing_extensions==4.12.2
validators==0.34.0
Werkzeug==3.0.3
wrapt==1.17.2
